[WIP] update mainloop and mainloop test
