+++ masr/mainloop.py
@@ -1,7 +1,7 @@
 import asyncio
 
 # from typing import Dict
-import uuid
+# import uuid
 
 
 class Promise:
@@ -28,54 +28,6 @@ class TaskPromise:
         self._promise_handler(self._promise)
 
 
-async def send(url):
-    await asyncio.sleep(1)
-    return "SUCCESS"
-
-
-class Env:
-    def __init__(self):
-        self.url = "http://mas.url"
-        self._success = False
-
-    def init(self, task_config):
-        pass
-
-    async def run(self):
-        return not self._success
-
-    async def send(self, url):
-        await asyncio.sleep(1)
-        return "SUCCESS"
-
-    async def check(self, promise):
-        await asyncio.sleep(1)
-        return Report(status="SUCCESS")  # Mocked success for simplicity
-
-    def set_success_flag(self, tid):
-        self._success = True
-
-
-class Mas:
-    def __init__(self):
-        self.url = "http://env.url"
-
-    def init(self, task_config):
-        pass
-
-    async def send(self, url):
-        await asyncio.sleep(1)
-        return "SUCCESS"
-
-    async def handel_failure(self, report):  # start again if fail
-        print("Handeling failure", report)
-
-
-class Report:
-    def __init__(self, status):
-        self.status = status
-
-
 async def mainloop(task_config, env, mas):
     env.init(task_config)
     mas.init(task_config)
@@ -98,11 +50,3 @@ async def mainloop(task_config, env, mas):
         else:
             await mas.handel_failure(report)
             await asyncio.sleep(1)
-
-
-if __name__ == "__main__":
-    env = Env()
-    mas = Mas()
-    task_config = {"task_id": uuid.uuid4()}
-
-    asyncio.run(mainloop(task_config, env, mas))
