+++ masr/models/interface.py
@@ -1,6 +1,6 @@
 # import uuid
-from uuid import UUID
 from typing import Dict
+from uuid import UUID
 
 from pydantic import BaseModel
