+++ tests/mainloop/test_mainloop.py
@@ -1,11 +1,62 @@
 import pytest
-from masr.mainloop import Mas, Env, Report, mainloop
+from masr.mainloop import mainloop
+import asyncio
 import uuid
+
+
+async def send(url):
+    await asyncio.sleep(1)
+    return "SUCCESS"
+
+
+class Env:
+    def __init__(self):
+        self.url = "http://mas.url"
+        self._success = False
+
+    def init(self, task_config):
+        pass
+
+    async def run(self):
+        return not self._success
+
+    async def send(self, url):
+        await asyncio.sleep(1)
+        return "SUCCESS"
+
+    async def check(self, promise):
+        await asyncio.sleep(1)
+        return Report(status="SUCCESS")  # Mocked success for simplicity
+
+    def set_success_flag(self, tid):
+        self._success = True
+
+
+class Mas:
+    def __init__(self):
+        self.url = "http://env.url"
+
+    def init(self, task_config):
+        pass
+
+    async def send(self, url):
+        await asyncio.sleep(1)
+        return "SUCCESS"
+
+    async def handel_failure(self, report):  # start again if fail
+        print("Handeling failure", report)
+
+
+class Report:
+    def __init__(self, status):
+        self.status = status
 
 
 @pytest.fixture
 def mock_task_config():
-    return {"task_id": uuid.uuid4()}
+    return {
+        uuid.uuid4(): {"demand": "task123", "report": "good", "src": "code"}
+    }
 
 
 @pytest.fixture
